requirements.txt
flask
flask-cors
gunicorn
python-dotenv
requests
azure-ai-formrecognizer
azure-core
PyMuPDF
